Move testing addresses into testing/ folder for collocation
